requirements.txt
anyio==4.9.0
asgiref==3.8.1
certifi==2025.1.31
channels==4.2.2
charset-normalizer==3.4.1
Django==5.2
djangorestframework==3.16.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
pillow==11.1.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-telegram-bot==22.0
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.13.1
urllib3==2.3.0
